requirements_cleaned.txt
# LimeClicks Requirements - Essential Dependencies Only
# Python 3.11+ Required
# Last updated: 2025-08-27

# Core Django & Framework
Django==5.2.5
asgiref==3.8.1
sqlparse==0.5.3
django-environ==0.12.0
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.10
dj-database-url==3.0.1

# Celery & Task Queue
celery==5.4.0
django-celery-beat==2.8.1
redis==6.2.0
django-redis==6.0.0
amqp==5.3.1
billiard==4.2.1
kombu==5.4.2
vine==5.1.0
python-crontab==3.3.0
cron-descriptor==1.4.5
django-timezone-field==7.1

# Authentication & Security
django-allauth==65.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cryptography==45.0.6
PyJWT==2.10.1

# Storage & Media
boto3==1.40.13
botocore==1.40.13
django-storages==1.14.6
Collectfasta==3.3.0
pillow==11.3.0
s3transfer==0.13.1
jmespath==1.0.1

# Email Services
django-anymail==13.0
brevo-python==1.2.0

# Admin & UI
django-unfold==0.64.2
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
crispy-tailwind==1.0.3
django-htmx==1.23.2

# Rich Text Editor
django-ckeditor-5==0.2.18
django-js-asset==3.1.2

# Utilities
django-model-utils==5.0.0
django-simple-history==3.10.1
python-slugify==8.0.4
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.1
humanize==4.12.3

# HTTP & Networking
requests==2.32.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
beautifulsoup4==4.13.5
soupsieve==2.7
websockets==15.0.1

# Server & ASGI
gunicorn==23.0.0
uvicorn==0.35.0
honcho==2.0.0
watchfiles==1.1.0

# CLI & Terminal
click==8.1.8
colorama==0.4.6
tqdm==4.67.1

# Core Dependencies
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyYAML==6.0.2
six==1.17.0
text-unidecode==1.3
typing-extensions==4.14.1
setuptools==80.9.0
wheel==0.45.1